[refactor] 검색 프래그먼트 챗봇 프래그먼트로 변경
