+++ android-app/app/src/main/res/layout/fragment_chatbot.xml
